src/event_stream: relocate inline tests to tests.rs
event_stream/ is a directory module. Extract 260-LOC inline test
block from mod.rs to sibling tests.rs.

After: event_stream/mod.rs drops from 943 → 686 LOC.

cargo build --release: clean. cargo test --release: 878 passed.
